copy push

+++ vectors_vs_seq_platform.py
@@ -155,7 +155,7 @@ class TransferDB(object):
         for id in self.dirty_transfers_dict.keys():
             try: 
                 
-                diff = float(1 - (int(self.clean_transfers_dict[id])/int(self.dirty_transfers_dict[id])))
+                diff = float(1 - float((int(self.clean_transfers_dict[id])/int(self.dirty_transfers_dict[id]))))
                 print "\t".join([id,str(self.dirty_transfers_dict[id]),str(self.clean_transfers_dict[id]),str(diff),self.dirty_seq_platform[id]])
             except KeyError:
                 diff = 1
